vendor_invoice_service/requirements.txt
fastapi
uvicorn
huggingface-hub
transformers
python-multipart
pydantic
requests
sqlalchemy
psycopg2-binary
python-dotenv
llama-cpp-python
pytesseract
pillow
